[#55] Move interface into a separate module
Resolves #55
