+++ test/Spec.hs
@@ -4,13 +4,13 @@ import Test.Hspec (hspec)
 
 import Test.Extensions (getExtensionsSpec)
 import Test.Extensions.Cabal (cabalSpec)
-import Test.Extensions.OnOff (onOffSpec)
-import Test.Extensions.Parser (parserSpec)
+import Test.Extensions.Module (moduleParserSpec)
+import Test.Extensions.Types (onOffSpec)
 
 
 main :: IO ()
 main = hspec $ do
     onOffSpec
     cabalSpec
-    parserSpec
+    moduleParserSpec
     getExtensionsSpec
